requirements.txt
autopep8==1.6.0
geopandas==0.10.2
pika==1.1.0
python-dotenv==0.10.5
requests==2.26.0
